Move plya_node to play direcotry
